Write the studio shell once instead of once per width
Codex found two defects in the desktop block, and both were real. Reproduced
before fixing, at 850px with a transcript stuffed past the window's height:

  850px  scrollerOverflowY: "visible"  pageScroll: 0  lastTurnBottom: 1692 (vh 900)
  850px  app-update overlapsComposer: true
  1440px app-update overlapsComposer: true

The first is the 801-900 band. A legacy block running to 900px turns the
transcript's scroller off — it was the one-scroller compromise from when the
page still scrolled — and the desktop shell takes the page's scroll away
without turning it back on. A long conversation ended 792px below the bottom of
the window with nothing to drag.

The second is the install nudge and the update bar, fixed at z-index 900,
sitting on the composer with no scroll left to escape them. That is the exact
bug #386 fixed for the phone, and the lesson written into AGENTS.md a commit
later.

Both have the same cause, which is the thing worth fixing: the desktop block
re-derived the shell from scratch instead of sharing it, and left behind the
two rules that only matter in states a screenshot does not reach. So the shell
now lives once, outside every media query — window-owning, footer-hiding,
scroller-owning, banner-reflowing — and the media blocks carry only what
differs: edge-to-edge and decluttered below 800, a rail and a capped column
above it.

Re-measured at 390, 780, 850, 1180 and 1440: page scroll 0, transcript scroller
live, neither bar over the composer, send button hittable at every one.

The e2e guard ran at 390 only, which is why it caught neither. It runs at four
widths now, asserts the transcript keeps a scroller once the page has none, and
is renamed studio-shell — it stopped being about phones.
